requirements.txt


joblib==1.2.0

keras==2.12.0

lightgbm==3.3.5

matplotlib==3.7.1
matplotlib-inline==0.1.6

numpy==1.24.2
openpyxl==3.1.2

pandas==1.5.3

Pillow==9.4.0
pipenv==2023.3.20

plotly==5.13.1



scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2


sns==0.1

streamlit==1.20.0



